+++ Core/TimberApi/Versions.cs
@@ -5,7 +5,7 @@ namespace TimberApi
 {
     public class Versions
     {
-        public static readonly Version GameVersion = Version.Parse(Timberborn.Versioning.Versions.VersionNumber);
+        public static readonly Version GameVersion = Version.Parse(Timberborn.Versioning.Versions.CurrentGameVersion.Numeric);
 
         public static Version TimberApiVersion { get; } =
             Version.Parse("TIMBER_API_VERSION_PLACEHOLDER".ReplacePlaceholderWithFakeVersion("1.0.0.0"));